+++ UCR.Plugins/Remapper/AxesToAxes.cs
@@ -57,9 +57,9 @@ namespace HidWizards.UCR.Plugins.Remapper
             _linearSenstitivityScaleFactor = ((double)Sensitivity / 100);
         }
 
-        public override void Update(params long[] values)
+        public override void Update(params short[] values)
         {
-            var outputValues = new long[] {values[0], values[1]};
+            var outputValues = new short[] {values[0], values[1]};
             if (DeadZone != 0)
             {
                 if (CircularDz)
@@ -77,8 +77,8 @@ namespace HidWizards.UCR.Plugins.Remapper
             {
                 if (Linear)
                 {
-                    outputValues[0] = (long)(outputValues[0] * _linearSenstitivityScaleFactor);
-                    outputValues[1] = (long)(outputValues[1] * _linearSenstitivityScaleFactor);
+                    outputValues[0] = (short) (outputValues[0] * _linearSenstitivityScaleFactor);
+                    outputValues[1] = (short) (outputValues[1] * _linearSenstitivityScaleFactor);
                 }
                 else
                 {
